+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- (gebaeude) add new bank building to increase the deposit limit
+
 ### Fixed
 
 - fix some minor inconsistencies for `Forschung`-columns in `mitglieder` database table